Add macro LOGURU_WITH_FILEABS
if LOGURU_WITH_FILEABS is defined, a check of file change
(registered by ``add_file'') will be done on every call to ``file_log''.
If file is moved or inode changes, file is reopened using the same
``FileMode'' as is done by ``add_file''.

Such a scheme is useful if you have a daemon program that moves the
log file every 24 hours and expects new file to be created.