Check for __GLIBC__ before enabling STACKTRACES
`LOGURU_STACKTRACES` depends on `execinfo.h`. However, not only is `execinfo.h` unavailable on Android, it's strictly a glibc extension, so everything that is POSIX, but not using glibc is not going to compile loguru (except, currently, Adroid).
Like I said in #86, `__GLIBC__` evaluates to the major version number of glibc. Alternatives are `__GLIBC_MINOR__` and `__GNU_LIBRARY__`.

Fixes #86