Check isatty() to decide if colors are supported
Currently, the type of terminal ($TERM) is checked in s_terminal_has_color() to decide if colors are supported. However, the stderr can be redirected to a file or another process which doesn't support colors. To check if stderr is being received by a terminal, at least the isatty() POSIX function must be used.