Virtualize EcEntryBase's destructor so it's called by derived classes
Fixes an error flagged by LGTM:

Derived classes from a base class without a
virtual destructor will only call the
destructor of the type of the pointer being
deleted and not the base-class.

This can cause a defect if the pointer type is
a base type while the object instance is a
derived type.