Fix clang on windows warnings (#180)
* Fix warning suppression on Windows clang

When building loguru under Windows clang, _MSC_VER is defined, but it
uses clang-style warning suppression flags. Fix that by checking for
__GNUC__ and __clang__ first, then checking for _MSC_VER.

Also fix the lack of a `#pragma GCC diagnostic pop` for people who
include loguru.cpp.

* Fix unsigned/signed comparison

`snprintf` can return a negative value to signify that the buffer is not
large enough to contain the bytes you attempted to write, so check for a
positive value before advancing `pos` -- otherwise a later `snprintf`
might overwrite bytes we've already written to the buffer.