Make the use of locale optional via LOGURU_USE_LOCALE Application developers can use the established LOGURU_USE_ mechanism to toggle the use of LOCALE:
"#define LOGURU_USE_LOCALE 1" or 0
or nothing, which will use LOCALE by default.

Application developers may want to avoid calls into the
the host's locale system, which can cause problems if the
external application is expecting to use a narrow
character set but gets a wide character set (and
vice-versa).

Co-authored-by: llm <dl.soluz@gmx.net>
Co-authored-by: kcgen <kcgen@users.noreply.github.com>